More file rename, now to make applet to upload items
